Characterize connection details query state
